feat: add ripgrep binding
